@@ -24,7 +24,7 @@
   "homepage": "https://github.com/awslabs/aws-crt-nodejs#readme",
   "devDependencies": {
     "@types/node": "^10.17.17",
-    "typescript": "^3.8.3"
+    "typescript": "^4.7.4"
   },
   "dependencies": {
     "aws-crt": "file:../../",
