@@ -29,7 +29,7 @@ export class AwsIotMqttConnectionConfigBuilder {
             client_id: '',
             host_name: '',
             socket_options: new SocketOptions(),
-            port: 8883,
+            port: 443,
             clean_session: false,
             keep_alive: undefined,
             will: undefined,
@@ -248,13 +248,13 @@ export class AwsIotMqttConnectionConfigBuilder {
      *                 check to see if a username has already been set (via WithUsername function). If no
      *                 username is set then no username will be passed with the MQTT connection.
      * @param authorizer_name The name of the custom authorizer. If an empty string is passed, then
-     *                       'x-amz-customauthorizer-name' will not be added with the MQTT connection.
+     *                       'x-amz-customauthorizer-name' will not be added with the MQTT connection.  It is strongly
+     *                       recommended to URL-encode this value; the SDK will not do so for you.
      * @param authorizer_signature The signature of the custom authorizer. If an empty string is passed, then
      *                            'x-amz-customauthorizer-signature' will not be added with the MQTT connection.
      *                            The signature must be based on the private key associated with the custom authorizer.
      *                            The signature must be base64 encoded.
-     *                            Required if the custom authorizer has signing enabled.  It is strongly suggested to URL-encode
-     *                            this value; the SDK will not do so for you.
+     *                            Required if the custom authorizer has signing enabled.
      * @param password The password to use with the custom authorizer. If null is passed, then no password will
      *                 be set.
      * @param token_key_name Key used to extract the custom authorizer token from MQTT username query-string properties.
@@ -265,8 +265,9 @@ export class AwsIotMqttConnectionConfigBuilder {
      *                    key associated with the custom authorizer and the result placed in the authorizer_signature argument.
      */
     with_custom_authorizer(username : string, authorizer_name : string, authorizer_signature : string, password : string, token_key_name? : string, token_value? : string) {
+        let uri_encoded_signature = iot_shared.canonicalizeCustomAuthTokenSignature(authorizer_signature);
         let username_string = iot_shared.populate_username_string_with_custom_authorizer(
-            "", username, authorizer_name, authorizer_signature, this.params.username, token_key_name, token_value);
+            "", username, authorizer_name, uri_encoded_signature, this.params.username, token_key_name, token_value);
         this.params.username = username_string;
         this.params.password = password;
         // Tells the websocket connection we are using a custom authorizer
