@@ -16,7 +16,7 @@ const CRuntimeType = Object.freeze({
 });
 
 function getCRuntime() {
-    if(platform() !== "linux") {
+    if (platform() !== "linux") {
         return CRuntimeType.NON_LINUX;
     }
 
@@ -88,5 +88,15 @@ if (binding == undefined) {
     throw new Error("AWS CRT binary not present in any of the following locations:\n\t" + search_paths.join('\n\t'));
 }
 
+import crt_native from "./binding"
+/** Electron will shutdown the node process on exit, which causes the threadsafe function to segfault. To prevent
+  * the segfault we disable the threadsafe function on node process exit. */
+if (process.versions.hasOwnProperty('electron')) {
+    process.on('exit', function () {
+        crt_native.disable_threadsafe_function();
+    });
+}
+
+
 export default binding;
-export { CRuntimeType , cRuntime };
+export { CRuntimeType, cRuntime };
