@@ -614,8 +614,18 @@ export class MqttClientConnection extends NativeResourceMixin(BufferedEventEmitt
      *
      * @group Node-only
      */
-    getQueueStatistics(): ConnectionStatistics {
+    getOperationalStatistics(): ConnectionStatistics {
         return crt_native.mqtt_client_connection_get_queue_statistics(this.native_handle());
+    }
+
+    /**
+     * Queries a small set of numerical statistics about the current state of the connection's operation queue
+     * @deprecated use getOperationalStatistics instead
+     *
+     * @group Node-only
+     */
+    getQueueStatistics(): ConnectionStatistics {
+        return this.getOperationalStatistics();
     }
 
     // Wrap a promise rejection with a function that will also emit the error as an event
