@@ -25,6 +25,7 @@
 #include <aws/common/logging.h>
 #include <aws/common/mutex.h>
 #include <aws/common/ref_count.h>
+#include <aws/common/rw_lock.h>
 #include <aws/common/system_info.h>
 
 #include <aws/event-stream/event_stream.h>
@@ -46,6 +47,16 @@
 AWS_STATIC_ASSERT(NAPI_VERSION >= 4);
 
 #define AWS_DEFINE_ERROR_INFO_CRT_NODEJS(CODE, STR) AWS_DEFINE_ERROR_INFO(CODE, STR, "aws-crt-nodejs")
+
+/* TODO:
+ * Hardcoded enum value for `napi_no_external_buffers_allowed`.
+ * The enum `napi_no_external_buffers_allowed` is introduced in node21 and backport
+ * to node 14.21.2, 16.19.0, 18.13.0.
+ * Use `napi_no_external_buffers_allowed` for external buffer related changes after bump to node 21 */
+#define NAPI_NO_EXTERNAL_BUFFER_ENUM_VALUE 22
+
+static bool s_tsfn_enabled = false;
+static struct aws_rw_lock s_tsfn_lock;
 
 /* clang-format off */
 static struct aws_error_info s_errors[] = {
@@ -268,9 +279,10 @@ int aws_napi_attach_object_property_binary_as_finalizable_external(
     }
 
     napi_value napi_binary = NULL;
+
     AWS_NAPI_ENSURE(
         env,
-        napi_create_external_arraybuffer(
+        aws_napi_create_external_arraybuffer(
             env,
             data_buffer->buffer,
             data_buffer->len,
@@ -742,6 +754,28 @@ int aws_napi_get_property_array_size(
     return AWS_OP_SUCCESS;
 }
 
+void s_aws_enable_threadsafe_function(void) {
+    aws_rw_lock_wlock(&s_tsfn_lock);
+    s_tsfn_enabled = true;
+    aws_rw_lock_wunlock(&s_tsfn_lock);
+}
+
+void s_aws_disable_threadsafe_function(void) {
+    aws_rw_lock_wlock(&s_tsfn_lock);
+    s_tsfn_enabled = false;
+    aws_rw_lock_wunlock(&s_tsfn_lock);
+}
+
+napi_value aws_napi_disable_threadsafe_function(napi_env env, napi_callback_info info) {
+    (void)info;
+    if (env == NULL) {
+        aws_raise_error(AWS_CRT_NODEJS_ERROR_THREADSAFE_FUNCTION_NULL_NAPI_ENV);
+        return NULL;
+    }
+    s_aws_disable_threadsafe_function();
+    return NULL;
+}
+
 void aws_napi_throw_last_error(napi_env env) {
     const int error_code = aws_last_error();
     napi_throw_error(env, aws_error_str(error_code), aws_error_debug_str(error_code));
@@ -831,7 +865,6 @@ const char *aws_napi_status_to_str(napi_status status) {
         case napi_callback_scope_mismatch:
             reason = "napi_callback_scope_mismatch";
             break;
-#if NAPI_VERSION >= 3
         case napi_queue_full:
             reason = "napi_queue_full";
             break;
@@ -841,7 +874,9 @@ const char *aws_napi_status_to_str(napi_status status) {
         case napi_bigint_expected:
             reason = "napi_bigint_expected";
             break;
-#endif
+        case NAPI_NO_EXTERNAL_BUFFER_ENUM_VALUE:
+            reason = "napi_no_external_buffers_allowed";
+            break;
     }
     return reason;
 }
@@ -931,6 +966,40 @@ static void s_handle_failed_callback(napi_env env, napi_value function, napi_sta
     }
 }
 
+napi_status aws_napi_create_external_arraybuffer(
+    napi_env env,
+    void *external_data,
+    size_t byte_length,
+    napi_finalize finalize_cb,
+    void *finalize_hint,
+    napi_value *result) {
+
+    napi_status external_buffer_status =
+        napi_create_external_arraybuffer(env, external_data, byte_length, finalize_cb, finalize_hint, result);
+
+    if (external_buffer_status == NAPI_NO_EXTERNAL_BUFFER_ENUM_VALUE) {
+
+        // The external buffer is disabled, manually copy the external_data into Node
+        void *napi_buf_data = NULL;
+        napi_status create_arraybuffer_status = napi_create_arraybuffer(env, byte_length, &napi_buf_data, result);
+
+        if (create_arraybuffer_status != napi_ok) {
+            AWS_NAPI_LOGF_ERROR(
+                "napi_create_arraybuffer (in aws_napi_create_external_arraybuffer) failed with : %s",
+                aws_napi_status_to_str(create_arraybuffer_status));
+            return create_arraybuffer_status;
+        }
+
+        memcpy(napi_buf_data, external_data, byte_length);
+
+        // As the data has been copied into the Node, invoke the finalize callback to make sure the
+        // data is released.
+        finalize_cb(env, finalize_hint, finalize_hint);
+    }
+
+    return napi_ok;
+}
+
 napi_status aws_napi_dispatch_threadsafe_function(
     napi_env env,
     napi_threadsafe_function tsfn,
@@ -939,19 +1008,25 @@ napi_status aws_napi_dispatch_threadsafe_function(
     size_t argc,
     napi_value *argv) {
 
-    napi_status call_status = napi_ok;
-    if (!this_ptr) {
-        AWS_NAPI_ENSURE(env, napi_get_undefined(env, &this_ptr));
+    aws_rw_lock_rlock(&s_tsfn_lock);
+    napi_status result = napi_ok;
+    if (s_tsfn_enabled) {
+        napi_status call_status = napi_ok;
+        if (!this_ptr) {
+            AWS_NAPI_ENSURE(env, napi_get_undefined(env, &this_ptr));
+        }
+        AWS_NAPI_CALL(env, napi_call_function(env, this_ptr, function, argc, argv, NULL), {
+            call_status = status;
+            s_handle_failed_callback(env, function, status);
+        });
+        /* main thread can exit now */
+        napi_unref_threadsafe_function(env, tsfn);
+        /* Must always decrement the ref count, or the function will be pinned */
+        napi_status release_status = napi_release_threadsafe_function(tsfn, napi_tsfn_release);
+        result = (call_status != napi_ok) ? call_status : release_status;
     }
-    AWS_NAPI_CALL(env, napi_call_function(env, this_ptr, function, argc, argv, NULL), {
-        call_status = status;
-        s_handle_failed_callback(env, function, status);
-    });
-    /* main thread can exit now */
-    napi_unref_threadsafe_function(env, tsfn);
-    /* Must always decrement the ref count, or the function will be pinned */
-    napi_status release_status = napi_release_threadsafe_function(tsfn, napi_tsfn_release);
-    return (call_status != napi_ok) ? call_status : release_status;
+    aws_rw_lock_runlock(&s_tsfn_lock);
+    return result;
 }
 
 napi_status aws_napi_create_threadsafe_function(
@@ -982,30 +1057,45 @@ napi_status aws_napi_create_threadsafe_function(
 napi_status aws_napi_release_threadsafe_function(
     napi_threadsafe_function function,
     napi_threadsafe_function_release_mode mode) {
-    if (function) {
-        return napi_release_threadsafe_function(function, mode);
+    napi_status result = napi_ok;
+    aws_rw_lock_rlock(&s_tsfn_lock);
+    if (s_tsfn_enabled && function) {
+        result = napi_release_threadsafe_function(function, mode);
     }
-    return napi_ok;
+    aws_rw_lock_runlock(&s_tsfn_lock);
+    return result;
 }
 
 napi_status aws_napi_acquire_threadsafe_function(napi_threadsafe_function function) {
-    if (function) {
-        return napi_acquire_threadsafe_function(function);
+    napi_status result = napi_ok;
+    aws_rw_lock_rlock(&s_tsfn_lock);
+    if (s_tsfn_enabled && function) {
+        result = napi_acquire_threadsafe_function(function);
     }
-    return napi_ok;
+    aws_rw_lock_runlock(&s_tsfn_lock);
+    return result;
 }
 
 napi_status aws_napi_unref_threadsafe_function(napi_env env, napi_threadsafe_function function) {
-    if (function) {
-        return napi_unref_threadsafe_function(env, function);
+    napi_status result = napi_ok;
+    aws_rw_lock_rlock(&s_tsfn_lock);
+    if (s_tsfn_enabled && function) {
+        result = napi_unref_threadsafe_function(env, function);
     }
-    return napi_ok;
+    aws_rw_lock_runlock(&s_tsfn_lock);
+    return result;
 }
 
 napi_status aws_napi_queue_threadsafe_function(napi_threadsafe_function function, void *user_data) {
-    /* increase the ref count, gets decreased when the call completes */
-    AWS_NAPI_ENSURE(NULL, napi_acquire_threadsafe_function(function));
-    return napi_call_threadsafe_function(function, user_data, napi_tsfn_nonblocking);
+    napi_status result = napi_ok;
+    aws_rw_lock_rlock(&s_tsfn_lock);
+    if (s_tsfn_enabled && function) {
+        /* increase the ref count, gets decreased when the call completes */
+        AWS_NAPI_ENSURE(NULL, napi_acquire_threadsafe_function(function));
+        result = napi_call_threadsafe_function(function, user_data, napi_tsfn_nonblocking);
+    }
+    aws_rw_lock_runlock(&s_tsfn_lock);
+    return result;
 }
 
 AWS_STATIC_STRING_FROM_LITERAL(s_mem_tracing_env_var, "AWS_CRT_MEMORY_TRACING");
@@ -1109,6 +1199,7 @@ static void s_napi_context_finalize(napi_env env, void *user_data, void *finaliz
     --s_module_initialize_count;
 
     if (s_module_initialize_count == 0) {
+
         aws_client_bootstrap_release(s_default_client_bootstrap);
         s_default_client_bootstrap = NULL;
 
@@ -1127,6 +1218,8 @@ static void s_napi_context_finalize(napi_env env, void *user_data, void *finaliz
         aws_mqtt_library_clean_up();
 
         s_uninstall_crash_handler();
+        // clean up threadsafe function lock
+        aws_rw_lock_clean_up(&s_tsfn_lock);
     }
 
     struct aws_napi_context *ctx = user_data;
@@ -1142,7 +1235,6 @@ static void s_napi_context_finalize(napi_env env, void *user_data, void *finaliz
             aws_mem_tracer_destroy(ctx_allocator);
         }
     }
-
     aws_mutex_unlock(&s_module_lock);
 }
 
@@ -1194,6 +1286,9 @@ static bool s_create_and_register_function(
     struct aws_allocator *allocator = aws_napi_get_allocator();
 
     if (s_module_initialize_count == 0) {
+        aws_rw_lock_init(&s_tsfn_lock);
+        s_aws_enable_threadsafe_function();
+
         s_install_crash_handler();
 
         aws_mqtt_library_init(allocator);
@@ -1259,6 +1354,7 @@ static bool s_create_and_register_function(
     CREATE_AND_REGISTER_FN(native_memory_dump)
     CREATE_AND_REGISTER_FN(error_code_to_string)
     CREATE_AND_REGISTER_FN(error_code_to_name)
+    CREATE_AND_REGISTER_FN(disable_threadsafe_function)
 
     /* IO */
     CREATE_AND_REGISTER_FN(io_logging_enable)
