@@ -29,6 +29,7 @@ module.exports = {
         '@common/(.+)': '<rootDir>/lib/common/$1',
         '@awscrt/(.+)': '<rootDir>/lib/browser/$1',
         '@awscrt': '<rootDir>/lib/browser.ts',
-        '@test/(.+)': '<rootDir>/test/$1'
+        '@test/(.+)': '<rootDir>/test/$1',
+        "axios": "axios/dist/node/axios.cjs"
     }
 }
